+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `Nanook::Node#receive_minimum`.
 - Added `Nanook::Wallet#move_accounts`.
 - Added `Nanook::Wallet#remove_account`.
+- Added `Nanook::Wallet#search_pending`.
 - Added `Nanook::Block#account`.
 - Added `Nanook::Block#amount`.
 - Added `Nanook::Block#balance`.